Check if connected before applying automatic validations.
Real world need this solves:

* I needed to pre-compile my assets for production in an environment without
  access to the production database server.
* I enable "config.assets.initialize_on_precompile" in my application.rb
  to allow for this. Now assets will compile without initializing rails.
* Since I am running in production mode config.eager_load is enabled
  in my environments/production.rb file.

Without this patch the eager loading loads my models which calls
enforce_schema_rules which starts a database connection so it can scan
the table for validations.

With this patch enforce_schema_rules will only do it's scan if we already
have a database connection. In a normal web or console environment we will
have a connection since Rails will initialize.