Fix tests. Failing column length validation.
The message on column length validation changed. If not set it now returns
a hash allowing the "one" message to be completely different than the "other"
message (i.e. not "one").

New code checks 1, then it outputs the "one" message directly. Otherwise it
formats the "other" message as the old code did. If the message is a string
then most likely it is a custom message so use the old code.