Remove foreign key validation.
Why I like the idea of foreign key validation I found the feature
buggy. Since it is validating on the field and not the relationship
I often end up with a false positive. for example:

    class Post
      enforce_schema_rules
      has_many :comments
    end

    class Comment
      enforce_schema_rules
      belongs_to :post
    end

    post.comments.create! message: 'Foo'

This will give an error saying the comment could not be created because
it failed validation and a comment needs a post. The workaround is to
either ignore the field with `enforce_schema_rules expect: :post_id` or
by this sort of odd statement:

    post.comments.create! message: 'Foo', post: post

The following article indicate validating the relationship is the proper method:

    http://railsguides.net/belongs-to-and-presence-validation-rule1/

I don't really have the bandwidth to determine if this is the best way to and implement
and test the code but I wanted to update the gem so that the broken method is no longer
applied.