Remove usage of `number?` predicate method. Rails 5 compat.
In Rails 5, many of the predicate methods on the column object have
been removed including `numeric?`. There seems to be no replacement.

I thought about querying the columns `sql_type_metadata` attribute and
seeing if it includes the `Type::Helpers::Numeric` module but this
seemed to reach too much in Rails' impelementation details and would be
likely to break.

Instead I looked at all the types that include that module and they are:

* float
* decimal
* integer

`:integer` is excluded in the original implementation (not sure I understand
the reason). For now just am checking for `float` and `decimal`. This should
be fully compatible with older versions of Rails.